backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
speedtest-cli==2.1.3
python-dotenv==1.0.0
alembic==1.12.1
apscheduler==3.10.4
httpx==0.25.2
